# Optimized rect getters