--- a/src_c/rect.c
+++ b/src_c/rect.c
@@ -1835,7 +1835,13 @@ pg_rect_setcentery(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_gettopleft(pgRectObject *self, void *closure)
 {
-    return Py_BuildValue("(ii)", self->r.x, self->r.y);
+    PyObject *tuple = PyTuple_New(2);
+    if (!tuple) {
+        return NULL;
+    }
+    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x));
+    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y));
+    return tuple;
 }
 
 static int
@@ -1862,7 +1868,13 @@ pg_rect_settopleft(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_gettopright(pgRectObject *self, void *closure)
 {
-    return Py_BuildValue("(ii)", self->r.x + self->r.w, self->r.y);
+    PyObject *tuple = PyTuple_New(2);
+    if (!tuple) {
+        return NULL;
+    }
+    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x + self->r.w));
+    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y));
+    return tuple;
 }
 
 static int
@@ -1889,7 +1901,13 @@ pg_rect_settopright(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getbottomleft(pgRectObject *self, void *closure)
 {
-    return Py_BuildValue("(ii)", self->r.x, self->r.y + self->r.h);
+    PyObject *tuple = PyTuple_New(2);
+    if (!tuple) {
+        return NULL;
+    }
+    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x));
+    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y + self->r.h));
+    return tuple;
 }
 
 static int
@@ -1916,7 +1934,13 @@ pg_rect_setbottomleft(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getbottomright(pgRectObject *self, void *closure)
 {
-    return Py_BuildValue("(ii)", self->r.x + self->r.w, self->r.y + self->r.h);
+    PyObject *tuple = PyTuple_New(2);
+    if (!tuple) {
+        return NULL;
+    }
+    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x + self->r.w));
+    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y + self->r.h));
+    return tuple;
 }
 
 static int
@@ -1943,7 +1967,13 @@ pg_rect_setbottomright(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getmidtop(pgRectObject *self, void *closure)
 {
-    return Py_BuildValue("(ii)", self->r.x + (self->r.w >> 1), self->r.y);
+    PyObject *tuple = PyTuple_New(2);
+    if (!tuple) {
+        return NULL;
+    }
+    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x + (self->r.w >> 1)));
+    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y));
+    return tuple;
 }
 
 static int
@@ -1970,7 +2000,13 @@ pg_rect_setmidtop(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getmidleft(pgRectObject *self, void *closure)
 {
-    return Py_BuildValue("(ii)", self->r.x, self->r.y + (self->r.h >> 1));
+    PyObject *tuple = PyTuple_New(2);
+    if (!tuple) {
+        return NULL;
+    }
+    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x));
+    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y + (self->r.h >> 1)));
+    return tuple;
 }
 
 static int
@@ -1997,8 +2033,13 @@ pg_rect_setmidleft(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getmidbottom(pgRectObject *self, void *closure)
 {
-    return Py_BuildValue("(ii)", self->r.x + (self->r.w >> 1),
-                         self->r.y + self->r.h);
+    PyObject *tuple = PyTuple_New(2);
+    if (!tuple) {
+        return NULL;
+    }
+    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x + (self->r.w >> 1)));
+    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y + self->r.h));
+    return tuple;
 }
 
 static int
@@ -2025,8 +2066,13 @@ pg_rect_setmidbottom(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getmidright(pgRectObject *self, void *closure)
 {
-    return Py_BuildValue("(ii)", self->r.x + self->r.w,
-                         self->r.y + (self->r.h >> 1));
+    PyObject *tuple = PyTuple_New(2);
+    if (!tuple) {
+        return NULL;
+    }
+    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x + self->r.w));
+    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y + (self->r.h >> 1)));
+    return tuple;
 }
 
 static int
@@ -2053,8 +2099,13 @@ pg_rect_setmidright(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getcenter(pgRectObject *self, void *closure)
 {
-    return Py_BuildValue("(ii)", self->r.x + (self->r.w >> 1),
-                         self->r.y + (self->r.h >> 1));
+    PyObject *tuple = PyTuple_New(2);
+    if (!tuple) {
+        return NULL;
+    }
+    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x + (self->r.w >> 1)));
+    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y + (self->r.h >> 1)));
+    return tuple;
 }
 
 static int
@@ -2081,7 +2132,13 @@ pg_rect_setcenter(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getsize(pgRectObject *self, void *closure)
 {
-    return Py_BuildValue("(ii)", self->r.w, self->r.h);
+    PyObject *tuple = PyTuple_New(2);
+    if (!tuple) {
+        return NULL;
+    }
+    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.w));
+    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.h));
+    return tuple;
 }
 
 static int
@@ -2213,6 +2270,11 @@ MODINIT_DEFINE(rect)
        the module is not loaded.
     */
     import_pygame_base();
+    if (PyErr_Occurred()) {
+        return NULL;
+    }
+
+    import_pygame_math();
     if (PyErr_Occurred()) {
         return NULL;
     }