--- a/src_c/rect.c
+++ b/src_c/rect.c
@@ -2274,11 +2274,6 @@ MODINIT_DEFINE(rect)
         return NULL;
     }
 
-    import_pygame_math();
-    if (PyErr_Occurred()) {
-        return NULL;
-    }
-
     /* Create the module and add the functions */
     if (PyType_Ready(&pgRect_Type) < 0) {
         return NULL;