--- a/src_c/rect.c
+++ b/src_c/rect.c
@@ -37,6 +37,27 @@
 static PyTypeObject pgRect_Type;
 #define pgRect_Check(x) ((x)->ob_type == &pgRect_Type)
 
+#define PG_RETURN_TUPLE_FROMVALUES_INT(val1, val2) \
+    do {                                           \
+        PyObject *tup = PyTuple_New(2);            \
+        if (!tup) {                                \
+            return NULL;                           \
+        }                                          \
+        PyObject *tmp = PyLong_FromLong(val1);     \
+        if (!tmp) {                                \
+            Py_DECREF(tup);                        \
+            return NULL;                           \
+        }                                          \
+        PyTuple_SET_ITEM(tup, 0, tmp);             \
+        tmp = PyLong_FromLong(val2);               \
+        if (!tmp) {                                \
+            Py_DECREF(tup);                        \
+            return NULL;                           \
+        }                                          \
+        PyTuple_SET_ITEM(tup, 1, tmp);             \
+        return tup;                                \
+    } while (0)
+
 static int
 pg_rect_init(pgRectObject *, PyObject *, PyObject *);
 
@@ -1835,13 +1856,7 @@ pg_rect_setcentery(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_gettopleft(pgRectObject *self, void *closure)
 {
-    PyObject *tuple = PyTuple_New(2);
-    if (!tuple) {
-        return NULL;
-    }
-    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x));
-    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y));
-    return tuple;
+    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x, self->r.y);
 }
 
 static int
@@ -1868,13 +1883,7 @@ pg_rect_settopleft(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_gettopright(pgRectObject *self, void *closure)
 {
-    PyObject *tuple = PyTuple_New(2);
-    if (!tuple) {
-        return NULL;
-    }
-    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x + self->r.w));
-    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y));
-    return tuple;
+    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x + self->r.w, self->r.y);
 }
 
 static int
@@ -1901,13 +1910,7 @@ pg_rect_settopright(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getbottomleft(pgRectObject *self, void *closure)
 {
-    PyObject *tuple = PyTuple_New(2);
-    if (!tuple) {
-        return NULL;
-    }
-    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x));
-    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y + self->r.h));
-    return tuple;
+    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x, self->r.y + self->r.h);
 }
 
 static int
@@ -1934,13 +1937,8 @@ pg_rect_setbottomleft(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getbottomright(pgRectObject *self, void *closure)
 {
-    PyObject *tuple = PyTuple_New(2);
-    if (!tuple) {
-        return NULL;
-    }
-    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x + self->r.w));
-    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y + self->r.h));
-    return tuple;
+    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x + self->r.w,
+                                   self->r.y + self->r.h);
 }
 
 static int
@@ -1967,13 +1965,7 @@ pg_rect_setbottomright(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getmidtop(pgRectObject *self, void *closure)
 {
-    PyObject *tuple = PyTuple_New(2);
-    if (!tuple) {
-        return NULL;
-    }
-    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x + (self->r.w >> 1)));
-    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y));
-    return tuple;
+    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x + (self->r.w >> 1), self->r.y);
 }
 
 static int
@@ -2000,13 +1992,7 @@ pg_rect_setmidtop(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getmidleft(pgRectObject *self, void *closure)
 {
-    PyObject *tuple = PyTuple_New(2);
-    if (!tuple) {
-        return NULL;
-    }
-    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x));
-    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y + (self->r.h >> 1)));
-    return tuple;
+    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x, self->r.y + (self->r.h >> 1));
 }
 
 static int
@@ -2033,13 +2019,8 @@ pg_rect_setmidleft(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getmidbottom(pgRectObject *self, void *closure)
 {
-    PyObject *tuple = PyTuple_New(2);
-    if (!tuple) {
-        return NULL;
-    }
-    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x + (self->r.w >> 1)));
-    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y + self->r.h));
-    return tuple;
+    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x + (self->r.w >> 1),
+                                   self->r.y + self->r.h);
 }
 
 static int
@@ -2066,13 +2047,8 @@ pg_rect_setmidbottom(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getmidright(pgRectObject *self, void *closure)
 {
-    PyObject *tuple = PyTuple_New(2);
-    if (!tuple) {
-        return NULL;
-    }
-    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x + self->r.w));
-    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y + (self->r.h >> 1)));
-    return tuple;
+    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x + self->r.w,
+                                   self->r.y + (self->r.h >> 1));
 }
 
 static int
@@ -2099,13 +2075,8 @@ pg_rect_setmidright(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getcenter(pgRectObject *self, void *closure)
 {
-    PyObject *tuple = PyTuple_New(2);
-    if (!tuple) {
-        return NULL;
-    }
-    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.x + (self->r.w >> 1)));
-    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.y + (self->r.h >> 1)));
-    return tuple;
+    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x + (self->r.w >> 1),
+                                   self->r.y + (self->r.h >> 1));
 }
 
 static int
@@ -2132,13 +2103,7 @@ pg_rect_setcenter(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getsize(pgRectObject *self, void *closure)
 {
-    PyObject *tuple = PyTuple_New(2);
-    if (!tuple) {
-        return NULL;
-    }
-    PyTuple_SET_ITEM(tuple, 0, PyLong_FromLong(self->r.w));
-    PyTuple_SET_ITEM(tuple, 1, PyLong_FromLong(self->r.h));
-    return tuple;
+    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.w, self->r.h);
 }
 
 static int