--- a/src_c/rect.c
+++ b/src_c/rect.c
@@ -37,27 +37,6 @@
 static PyTypeObject pgRect_Type;
 #define pgRect_Check(x) ((x)->ob_type == &pgRect_Type)
 
-#define PG_RETURN_TUPLE_FROMVALUES_INT(val1, val2) \
-    do {                                           \
-        PyObject *tup = PyTuple_New(2);            \
-        if (!tup) {                                \
-            return NULL;                           \
-        }                                          \
-        PyObject *tmp = PyLong_FromLong(val1);     \
-        if (!tmp) {                                \
-            Py_DECREF(tup);                        \
-            return NULL;                           \
-        }                                          \
-        PyTuple_SET_ITEM(tup, 0, tmp);             \
-        tmp = PyLong_FromLong(val2);               \
-        if (!tmp) {                                \
-            Py_DECREF(tup);                        \
-            return NULL;                           \
-        }                                          \
-        PyTuple_SET_ITEM(tup, 1, tmp);             \
-        return tup;                                \
-    } while (0)
-
 static int
 pg_rect_init(pgRectObject *, PyObject *, PyObject *);
 
@@ -165,6 +144,31 @@ four_ints_from_obj(PyObject *obj, int *val1, int *val2, int *val3, int *val4)
     }
 
     return 1;
+}
+
+PyObject *
+pg_tuple_from_values_int(int val1, int val2)
+{
+    PyObject *tup = PyTuple_New(2);
+    if (!tup) {
+        return NULL;
+    }
+
+    PyObject *tmp = PyLong_FromLong(val1);
+    if (!tmp) {
+        Py_DECREF(tup);
+        return NULL;
+    }
+    PyTuple_SET_ITEM(tup, 0, tmp);
+
+    tmp = PyLong_FromLong(val2);
+    if (!tmp) {
+        Py_DECREF(tup);
+        return NULL;
+    }
+    PyTuple_SET_ITEM(tup, 1, tmp);
+
+    return tup;
 }
 
 static PyObject *
@@ -1856,7 +1860,7 @@ pg_rect_setcentery(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_gettopleft(pgRectObject *self, void *closure)
 {
-    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x, self->r.y);
+    return pg_tuple_from_values_int(self->r.x, self->r.y);
 }
 
 static int
@@ -1883,7 +1887,7 @@ pg_rect_settopleft(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_gettopright(pgRectObject *self, void *closure)
 {
-    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x + self->r.w, self->r.y);
+    return pg_tuple_from_values_int(self->r.x + self->r.w, self->r.y);
 }
 
 static int
@@ -1910,7 +1914,7 @@ pg_rect_settopright(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getbottomleft(pgRectObject *self, void *closure)
 {
-    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x, self->r.y + self->r.h);
+    return pg_tuple_from_values_int(self->r.x, self->r.y + self->r.h);
 }
 
 static int
@@ -1937,8 +1941,8 @@ pg_rect_setbottomleft(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getbottomright(pgRectObject *self, void *closure)
 {
-    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x + self->r.w,
-                                   self->r.y + self->r.h);
+    return pg_tuple_from_values_int(self->r.x + self->r.w,
+                                    self->r.y + self->r.h);
 }
 
 static int
@@ -1965,7 +1969,7 @@ pg_rect_setbottomright(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getmidtop(pgRectObject *self, void *closure)
 {
-    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x + (self->r.w >> 1), self->r.y);
+    return pg_tuple_from_values_int(self->r.x + (self->r.w >> 1), self->r.y);
 }
 
 static int
@@ -1992,7 +1996,7 @@ pg_rect_setmidtop(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getmidleft(pgRectObject *self, void *closure)
 {
-    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x, self->r.y + (self->r.h >> 1));
+    return pg_tuple_from_values_int(self->r.x, self->r.y + (self->r.h >> 1));
 }
 
 static int
@@ -2019,8 +2023,8 @@ pg_rect_setmidleft(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getmidbottom(pgRectObject *self, void *closure)
 {
-    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x + (self->r.w >> 1),
-                                   self->r.y + self->r.h);
+    return pg_tuple_from_values_int(self->r.x + (self->r.w >> 1),
+                                    self->r.y + self->r.h);
 }
 
 static int
@@ -2047,8 +2051,8 @@ pg_rect_setmidbottom(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getmidright(pgRectObject *self, void *closure)
 {
-    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x + self->r.w,
-                                   self->r.y + (self->r.h >> 1));
+    return pg_tuple_from_values_int(self->r.x + self->r.w,
+                                    self->r.y + (self->r.h >> 1));
 }
 
 static int
@@ -2075,8 +2079,8 @@ pg_rect_setmidright(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getcenter(pgRectObject *self, void *closure)
 {
-    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.x + (self->r.w >> 1),
-                                   self->r.y + (self->r.h >> 1));
+    return pg_tuple_from_values_int(self->r.x + (self->r.w >> 1),
+                                    self->r.y + (self->r.h >> 1));
 }
 
 static int
@@ -2103,7 +2107,7 @@ pg_rect_setcenter(pgRectObject *self, PyObject *value, void *closure)
 static PyObject *
 pg_rect_getsize(pgRectObject *self, void *closure)
 {
-    PG_RETURN_TUPLE_FROMVALUES_INT(self->r.w, self->r.h);
+    return pg_tuple_from_values_int(self->r.w, self->r.h);
 }
 
 static int