@itzpr3d4t0r 2022-08-19
Yep sure, was thinking of adding a macro already